refactor(logic): complete Slice D orders/combat/diplomacy/turn purge (Refs #4090)
Port unique army/civilian/draft/combat/diplomacy/turn suites into owning Melos packages and delete orphans covered by domain peers; inventory in REFACTOR_TRACE.
