+++ packages/colonizethis_orders/test/orders/draft_orders_naval_test.dart
@@ -1,3 +1,4 @@
+// Ported from colonizethis_logic (Refs #4090 Slice D).
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
