+++ packages/colonizethis_world/test/world/capital_reassignment_after_combat_unique_ports_test.dart
@@ -1,6 +1,6 @@
+// Ported from colonizethis_logic (Refs #4090 Slice D).
 import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_logic/colonizethis_logic.dart';
-import 'package:colonizethis_world/src/world/capital_and_gp_fall.dart';
+import 'package:colonizethis_world/colonizethis_world.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
